# openapi.yaml
openapi: 3.0.2
info:
  title: Voluble API
  version: '1.0'
  description: >
    This is the OpenAPI specification for the Voluble messaging service API.

    Current API status:![OpenAPI Validity
    Status](http://validator.swagger.io/validator?url=https://raw.githubusercontent.com/calmcl1/voluble/master/openapi.json)
    
    Note: all responses are `JSEND`-compatible, which means that the general success or failure of a request can be determined from the `status` field in the HTTP response. The API data returned can be found in the `data` field in the response.
  contact:
    email: mail@callum-mclean.co.uk
    name: Cal McLean
servers:
  - url: 'https://voluble-poc.herokuapp.com/v1'
...
components:
  responses:
    '400':
      description: Data supplied is invalid
    '404':
      description: No object found with the specified ID
    '410':
      description: Object with specified ID has already been deleted
    '500':
      description: Internal server error
  schemas:
    Contact:
      type: object
      properties:
        id:
          type: string
          description: UUID for this Contact within Voluble
          example: c616d114-4f68-4dba-b976-cd5ae0aa359d
          readOnly: true
          nullable: false
          format: uuid
          pattern: >-
            ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}
        first_name:
          type: string
          description: The first name of the Contact
          example: John
          nullable: false
        surname:
          type: string
          description: The surname of the Contact
          example: Cleese
          nullable: false
        email_address:
          type: string
          description: >-
            The email address for the contact, that will be passed to Auth0 for
            authentication
          example: myname@emailprovider.com
          nullable: false
        phone_number:
          type: string
          description: >-
            The E164-formatted phone number for the Contact that will be used
            for fallback SMS messaging
          example: '+442071838750'
          nullable: false
        OrganizationId:
          type: string
          description: >-
            The UUID of the Organization within Voluble that this Contact
            belongs to
          example: c616d114-4f68-4dba-b976-cd5ae0aa359d
          nullable: false
        ServicechainId:
          type: string
          description: >-
            The UUID of the default servicechain that will be used to send
            messages to the Contact
          nullable: false
          example: c616d114-4f68-4dba-b976-cd5ae0aa359d
    Message:
      type: object
      properties:
        id:
          type: string
          description: The UUID of the Message within Voluble
          readOnly: true
          example: c616d114-4f68-4dba-b976-cd5ae0aa359d
          nullable: false
        body:
          type: string
          description: The text content of the Message that is transmitted
          example: Hi! Voluble has delivered this message to you!
          nullable: true
        ServicechainID:
          type: string
          description: >-
            The UUID of the of the Servicechain within Voluble to be used to
            send the message
          example: c616d114-4f68-4dba-b976-cd5ae0aa359d
          nullable: false
        contact:
          type: string
          nullable: false
          description: >-
            The UUID of the Contact related to this message (i.e. sender or
            recipient) within Voluble
          example: c616d114-4f68-4dba-b976-cd5ae0aa359d
        is_reply_to:
          type: string
          nullable: true
          description: >-
            The UUID of the Message that this Message is a reply to. NULL if
            this is message was not a reply to a previous message
          example: c616d114-4f68-4dba-b976-cd5ae0aa359d
        direction:
          type: string
          enum:
            - INBOUND
            - OUTBOUND
          description: >-
            Describes which direction the message was moving (i.e. who sent the
            message): `OUTBOUND` if the message was sent from Voluble to a
            Contact; `INBOUND` if the message was sent by a Contact to Voluble.
          nullable: false
          readOnly: true
        sent_time:
          type: number
          nullable: false
          description: The UNIX timestamp of when the message was sent.
          example: 1568548864
          readOnly: true
        message_state:
          type: string
          description: >-
            Describes the current state of the message - `SENT`, `FAILED`,
            `DELIVERED`, etc.
          nullable: false
          readOnly: true
          enum:
            - MSG_PENDING
            - MSG_SENDING
            - MSG_DELIVERED_SERVICE
            - MSG_DELIVERED_USER
            - MSG_READ
            - MSG_REPLIED
            - MSG_FAILED
            - MSG_ARRIVED
...
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'A JWT bearer token generated by Auth0, representing the User.'
paths:
  /contacts:
    get:
      tags:
        - Contacts
      summary: >-
        Returns a list of the first 100 Contacts available to the User in the
        Organization
      description: >-
        Returns a list of the first 100 Contacts available to the User in the
        Organization, starting at `offset`.
      security:
        - jwt: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contact'
                maxItems: 100
        '500':
          $ref: '#/components/responses/500'
    post:
      tags:
        - Contacts
      summary: Allows the creation of a new Contact
      security:
        - jwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Contact'
      responses:
        '201':
          description: 'OK: User created'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
          links:
            GetContactById:
              operationId: getContact
              parameters:
                contact_id: '$response.body#/id'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  '/contacts/{contact_id}':
    parameters:
      - name: contact_id
        in: path
        required: true
        description: The ID of the Contact to manipulate
        schema:
          type: string
    get:
      tags:
        - Contacts
      operationId: getContact
      summary: Retrieves a particular Contact
      security:
        - jwt: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
    put:
      tags:
        - Contacts
      operationId: modifyContact
      summary: Modifies a specific contact
      security:
        - jwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Contact'
      responses:
        '200':
          description: Contact modified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
          links:
            GetContactById:
              operationId: getContact
              parameters:
                contact_id: '$response.body#/id'
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
    delete:
      tags:
        - Contacts
      operationId: deleteContact
      summary: Deletes a particular contact from Voluble
      security:
        - jwt: []
      responses:
        '200':
          description: Contact deleted successfully
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '410':
          $ref: '#/components/responses/410'
        '500':
          $ref: '#/components/responses/500'
  /messages:
    get:
      tags:
        - Messages
      description: >-
        Returns a list of the first 100 Messages available to the User, starting
        at `offset`
      summary: Returns a list of the first 100 Messages available to the User
      security:
        - jwt: []
      parameters:
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
                maxItems: 100
        '500':
          $ref: '#/components/responses/500'
    post:
      tags:
        - Messages
      summary: Creates and sends a new Message
      security:
        - jwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Message'
      responses:
        '201':
          description: >-
            Message attempt successfully created; the created Message will be
            returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          links:
            GetMessageById:
              operationId: getMessage
              parameters:
                contact_id: '$response.body#/id'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  '/messages/{message_id}':
    parameters:
      - name: message_id
        description: The UUID of the Message within Voluble to manipulate
        required: true
        in: path
        schema:
          type: string
    get:
      tags:
        - Messages
      operationId: getMessage
      summary: Retrieves a specific Message
      security:
        - jwt: []
      responses:
        '200':
          description: Message retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'
...